Code Refactored.
Added event handling and decorator code.

+++ BookStore/src/main/java/org/myjee/cdi/practice/BookStore/Producers/InjectionProducer.java
@@ -32,7 +32,7 @@ public class InjectionProducer {
 
 	@Produces
 	@RandomNumber
-	public double random() {
+	public int random() {
 		return Math.abs(new Random().nextInt());
 	}
