+++ BookStore/src/main/java/org/myjee/cdi/practice/BookStore/numbergenerate/IssnGenerator.java
@@ -20,7 +20,7 @@ public class IssnGenerator extends NumberGenerator {
 
 	@Inject
 	@RandomNumber
-	private double randomNumber;
+	private int randomNumber;
 
 	@Loggable
 	public String generateNumber() {
